pim-engine/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Cache
redis==5.0.1
aiocache==0.12.2

# Utilities
pyyaml==6.0.1
python-multipart==0.0.6
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
psutil==5.9.8
aiohttp==3.9.3

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0

# LLM (optional for rule engine)
anthropic==0.8.0
langchain==0.3.18
langchain-community==0.3.18
langchain-openai==0.3.0
langchain-google-genai==2.0.7
google-generativeai==0.8.3
google-ai-generativelanguage==0.6.10
diskcache==5.6.3

# Development
pytest==7.4.0
pytest-asyncio==0.23.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3